requirements.txt
beautifulsoup4
google-cloud-datastore
